All search function is now working in payroll
